requirements.txt
altair==6.0.0
annotated-types==0.7.0
anthropic==0.85.0
anyio==4.12.1
asn1crypto==1.5.1
attrs==26.1.0
blinker==1.9.0
boto3==1.42.69
botocore==1.42.69
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cryptography==46.0.5
distro==1.9.0
docstring_parser==0.17.0
filelock==3.25.2
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
narwhals==2.18.0
numpy==1.26.4 
packaging==26.0
pandas==2.3.3
pillow==12.1.1
platformdirs==4.9.4
protobuf==6.33.6
pyarrow==23.0.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
PyJWT==2.12.1
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
s3transfer==0.16.0
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
snowflake-connector-python==4.3.0
sortedcontainers==2.4.0
streamlit==1.55.0
tenacity==9.1.4
toml==0.10.2
tomlkit==0.14.0
tornado==6.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
